refactor: move stream persistence into adapter
