+++ crates/aionui-conversation/src/stream_relay.rs
@@ -1,70 +1,25 @@
 use std::sync::Arc;
 
-use aionui_ai_agent::{
-    AgentSendError, AgentStreamEvent,
-    protocol::events::{
-        ThinkingEventData,
-        tool_call::{AcpToolCallSessionUpdateKind, AcpToolCallStatus, ToolCallStatus},
-    },
-};
+use aionui_ai_agent::{AgentSendError, AgentStreamEvent, protocol::events::ThinkingEventData};
 
 use crate::response_middleware::{ICronService, MessageMiddleware, MiddlewareResult};
-use aionui_api_types::{AgentErrorCode, ConversationRuntimeSummary, WebSocketMessage};
+use aionui_api_types::{AgentErrorCode, WebSocketMessage};
 use aionui_common::{ErrorChain, normalize_keys_to_snake_case, now_ms};
 
-use crate::runtime_persistence::{RuntimePersistenceCoordinator, RuntimeWriteKind};
+use crate::runtime_persistence::RuntimePersistenceCoordinator;
 use crate::runtime_state::ConversationRuntimeStateService;
 use crate::service::ConversationService;
-use aionui_db::models::MessageRow;
-use aionui_db::{DbError, IConversationRepository};
+use crate::stream_persistence::{
+    PersistedTextSegment, StreamPersistenceAdapter, TextSegmentState, ThinkingSegmentState,
+};
+use aionui_db::IConversationRepository;
 use aionui_realtime::EventBroadcaster;
 use serde_json::json;
 use tokio::sync::{broadcast, oneshot};
-use tracing::{debug, error, info, warn};
+use tracing::{debug, info, warn};
 
 /// Number of text chunks to accumulate before flushing to the database.
 const FLUSH_INTERVAL: u32 = 20;
-
-fn is_not_found(err: &DbError) -> bool {
-    matches!(err, DbError::NotFound(_))
-}
-
-fn is_foreign_key_constraint(err: &DbError) -> bool {
-    err.to_string().contains("FOREIGN KEY constraint failed")
-}
-
-fn is_deleted_during_stream_persistence(err: &DbError) -> bool {
-    is_not_found(err) || is_foreign_key_constraint(err)
-}
-
-fn log_persist_error(err: &DbError, message: &'static str) {
-    if is_deleted_during_stream_persistence(err) {
-        debug!(error = %ErrorChain(err), "{message}; conversation was likely deleted during stream finalization");
-    } else {
-        error!(error = %ErrorChain(err), "{message}");
-    }
-}
-
-#[derive(Debug, Clone)]
-struct TextSegmentState {
-    id: String,
-    buffer: String,
-    created_at: i64,
-    record_created: bool,
-    flush_counter: u32,
-}
-
-#[derive(Debug, Clone)]
-struct PersistedTextSegment {
-    id: String,
-}
-
-#[derive(Debug, Clone)]
-struct ThinkingSegmentState {
-    id: String,
-    buffer: String,
-    started_at: i64,
-}
 
 /// Result returned after a relay turn has fully drained and finalized.
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
@@ -112,11 +67,11 @@ pub struct StreamRelay {
     conversation_id: String,
     msg_id: String,
     user_id: String,
-    repo: Arc<dyn IConversationRepository>,
     broadcaster: Arc<dyn EventBroadcaster>,
     cron_service: Option<Arc<dyn ICronService>>,
     runtime_state: Option<Arc<ConversationRuntimeStateService>>,
     persistence: Option<RuntimePersistenceCoordinator>,
+    adapter: StreamPersistenceAdapter,
     complete_turn: bool,
 }
 
@@ -129,15 +84,16 @@ impl StreamRelay {
         broadcaster: Arc<dyn EventBroadcaster>,
         cron_service: Option<Arc<dyn ICronService>>,
     ) -> Self {
+        let adapter = StreamPersistenceAdapter::new(conversation_id.clone(), msg_id.clone(), repo, None);
         Self {
             conversation_id,
             msg_id,
             user_id,
-            repo,
             broadcaster,
             cron_service,
             runtime_state: None,
             persistence: None,
+            adapter,
             complete_turn: true,
         }
     }
@@ -148,7 +104,8 @@ impl StreamRelay {
     }
 
     pub fn with_persistence(mut self, persistence: RuntimePersistenceCoordinator) -> Self {
-        self.persistence = Some(persistence);
+        self.persistence = Some(persistence.clone());
+        self.adapter = self.adapter.with_persistence(persistence);
         self
     }
 
@@ -295,7 +252,7 @@ impl StreamRelay {
                             full_text_buffer.push_str(&data.content);
                             segment.flush_counter += 1;
                             if segment.flush_counter >= FLUSH_INTERVAL {
-                                self.flush_text_segment(segment).await;
+                                self.adapter.flush_text_segment(segment).await;
                                 segment.flush_counter = 0;
                             }
                         }
@@ -353,7 +310,7 @@ impl StreamRelay {
                                 self.finalize(&full_text_buffer, &text_segments, &event, terminal).await
                             };
                             if self.complete_turn && !deleting {
-                                Self::complete_conversation(&self.repo, &self.broadcaster, &self.conversation_id).await;
+                                self.adapter.complete_conversation(&self.broadcaster, None).await;
                             }
                             break outcome;
                         }
@@ -362,21 +319,21 @@ impl StreamRelay {
                             self.close_active_text_segment(&mut active_text, &mut text_segments, "finish")
                                 .await;
                             self.forward_to_websocket(&event);
-                            self.persist_tool_call(data).await;
+                            self.adapter.persist_tool_call(data).await;
                         }
                         AgentStreamEvent::AcpToolCall(data) => {
                             self.complete_active_thinking(&mut active_thinking).await;
                             self.close_active_text_segment(&mut active_text, &mut text_segments, "finish")
                                 .await;
                             self.forward_to_websocket(&event);
-                            self.persist_acp_tool_call(data).await;
+                            self.adapter.persist_acp_tool_call(data).await;
                         }
                         AgentStreamEvent::ToolGroup(entries) => {
                             self.complete_active_thinking(&mut active_thinking).await;
                             self.close_active_text_segment(&mut active_text, &mut text_segments, "finish")
                                 .await;
                             self.forward_to_websocket(&event);
-                            self.persist_tool_group(entries).await;
+                            self.adapter.persist_tool_group(entries).await;
                         }
                         _ => {
                             self.forward_to_websocket(&event);
@@ -415,7 +372,7 @@ impl StreamRelay {
                         .await
                     };
                     if self.complete_turn && !deleting {
-                        Self::complete_conversation(&self.repo, &self.broadcaster, &self.conversation_id).await;
+                        self.adapter.complete_conversation(&self.broadcaster, None).await;
                     }
                     break outcome;
                 }
@@ -430,12 +387,6 @@ impl StreamRelay {
         self.runtime_state
             .as_ref()
             .is_some_and(|state| state.is_deleting(&self.conversation_id))
-    }
-
-    fn allows_write(&self, kind: RuntimeWriteKind) -> bool {
-        self.persistence
-            .as_ref()
-            .is_none_or(|persistence| persistence.allows(&self.conversation_id, kind))
     }
 
     fn event_kind(event: &AgentStreamEvent) -> &'static str {
@@ -520,87 +471,6 @@ impl StreamRelay {
         self.broadcast_stream_payload(payload);
     }
 
-    /// Flush an active text segment to the database (create or update).
-    #[tracing::instrument(skip_all)]
-    async fn flush_text_segment(&self, segment: &mut TextSegmentState) {
-        if !self.allows_write(RuntimeWriteKind::AssistantTextFlush) {
-            return;
-        }
-        if segment.buffer.is_empty() {
-            return;
-        }
-
-        let content = json!({ "content": segment.buffer }).to_string();
-
-        if segment.record_created {
-            let update = aionui_db::MessageRowUpdate {
-                content: Some(content),
-                status: Some(Some("work".into())),
-                hidden: None,
-            };
-            if let Err(e) = self.repo.update_message(&segment.id, &update).await {
-                log_persist_error(&e, "Failed to update streaming text segment");
-            }
-        } else {
-            let row = MessageRow {
-                id: segment.id.clone(),
-                conversation_id: self.conversation_id.clone(),
-                msg_id: Some(segment.id.clone()),
-                r#type: "text".into(),
-                content,
-                position: Some("left".into()),
-                status: Some("work".into()),
-                hidden: false,
-                created_at: segment.created_at,
-            };
-            if let Err(e) = self.repo.insert_message(&row).await {
-                log_persist_error(&e, "Failed to create streaming text segment");
-            }
-            segment.record_created = true;
-        }
-    }
-
-    #[tracing::instrument(skip_all)]
-    async fn finalize_text_segment(&self, segment: TextSegmentState, status: &str) -> Option<PersistedTextSegment> {
-        if !self.allows_write(RuntimeWriteKind::AssistantTextFinalize) {
-            return None;
-        }
-        if segment.buffer.is_empty() {
-            return None;
-        }
-
-        let content = json!({ "content": segment.buffer }).to_string();
-        if segment.record_created {
-            let update = aionui_db::MessageRowUpdate {
-                content: Some(content),
-                status: Some(Some(status.to_owned())),
-                hidden: Some(false),
-            };
-            if let Err(e) = self.repo.update_message(&segment.id, &update).await {
-                log_persist_error(&e, "Failed to finalize text segment");
-                return None;
-            }
-        } else {
-            let row = MessageRow {
-                id: segment.id.clone(),
-                conversation_id: self.conversation_id.clone(),
-                msg_id: Some(segment.id.clone()),
-                r#type: "text".into(),
-                content,
-                position: Some("left".into()),
-                status: Some(status.to_owned()),
-                hidden: false,
-                created_at: segment.created_at,
-            };
-            if let Err(e) = self.repo.insert_message(&row).await {
-                log_persist_error(&e, "Failed to create finalized text segment");
-                return None;
-            }
-        }
-
-        Some(PersistedTextSegment { id: segment.id })
-    }
-
     /// Finalize assistant text on stream end and apply middleware rewrites.
     #[tracing::instrument(skip_all)]
     async fn finalize(
@@ -624,90 +494,19 @@ impl StreamRelay {
             let final_text = processed.message.trim().to_owned();
             let hidden = final_text.is_empty();
 
-            if let Some(primary_segment) = text_segments.first() {
-                if processed.message != text || hidden {
-                    if !self.allows_write(RuntimeWriteKind::TerminalFinalize) {
-                        return outcome;
-                    }
-                    let content = json!({ "content": final_text }).to_string();
-                    let update = aionui_db::MessageRowUpdate {
-                        content: Some(content),
-                        status: Some(Some(status.to_owned())),
-                        hidden: Some(hidden),
-                    };
-                    if let Err(e) = self.repo.update_message(&primary_segment.id, &update).await {
-                        log_persist_error(&e, "Failed to rewrite finalized text segment");
-                    }
-                    self.send_final_text_override(&primary_segment.id, &processed.message, hidden);
-
-                    for segment in text_segments.iter().skip(1) {
-                        let hide_update = aionui_db::MessageRowUpdate {
-                            content: None,
-                            status: Some(Some(status.to_owned())),
-                            hidden: Some(true),
-                        };
-                        if let Err(e) = self.repo.update_message(&segment.id, &hide_update).await {
-                            log_persist_error(&e, "Failed to hide superseded text segment");
-                        }
-                        self.send_final_text_override(&segment.id, "", true);
-                    }
-                } else {
-                    if !self.allows_write(RuntimeWriteKind::TerminalFinalize) {
-                        return outcome;
-                    }
-                    for segment in text_segments {
-                        let status_update = aionui_db::MessageRowUpdate {
-                            content: None,
-                            status: Some(Some(status.to_owned())),
-                            hidden: Some(false),
-                        };
-                        if let Err(e) = self.repo.update_message(&segment.id, &status_update).await {
-                            log_persist_error(&e, "Failed to finalize text segment status");
-                        }
-                    }
-                }
-            } else if !hidden {
-                if !self.allows_write(RuntimeWriteKind::TerminalFinalize) {
-                    return outcome;
-                }
-                let row = MessageRow {
-                    id: self.msg_id.clone(),
-                    conversation_id: self.conversation_id.clone(),
-                    msg_id: Some(self.msg_id.clone()),
-                    r#type: "text".into(),
-                    content: json!({ "content": final_text }).to_string(),
-                    position: Some("left".into()),
-                    status: Some(status.to_owned()),
-                    hidden: false,
-                    created_at: now_ms(),
-                };
-                if let Err(e) = self.repo.insert_message(&row).await {
-                    log_persist_error(&e, "Failed to create final fallback message");
-                }
+            let rewrite_segments = processed.message != text || hidden;
+            let overrides = self
+                .adapter
+                .persist_final_text(text_segments, status, &final_text, hidden, rewrite_segments)
+                .await;
+            for override_event in overrides {
+                self.send_final_text_override(&override_event.msg_id, &override_event.text, override_event.hidden);
             }
 
             self.send_system_responses(&processed.system_responses);
             outcome.system_responses = processed.system_responses;
         } else if let AgentStreamEvent::Error(data) = event {
-            if !self.allows_write(RuntimeWriteKind::TerminalFinalize) {
-                return outcome;
-            }
-            // No text accumulated but got an error — store error as tips message
-            let content = json!({ "content": &data.message, "type": "error", "error": &data }).to_string();
-            let row = MessageRow {
-                id: ConversationService::mint_msg_id(),
-                conversation_id: self.conversation_id.clone(),
-                msg_id: None,
-                r#type: "tips".into(),
-                content,
-                position: Some("left".into()),
-                status: Some("error".into()),
-                hidden: false,
-                created_at: now_ms(),
-            };
-            if let Err(e) = self.repo.insert_message(&row).await {
-                log_persist_error(&e, "Failed to store error message");
-            }
+            self.adapter.persist_error_tip(data).await;
         }
 
         outcome
@@ -720,32 +519,7 @@ impl StreamRelay {
         };
         let duration_ms = (now_ms() - segment.started_at).max(0);
         self.send_thinking_done(&segment.id, duration_ms as u64);
-        if segment.buffer.is_empty() {
-            return;
-        }
-        if !self.allows_write(RuntimeWriteKind::AssistantThinkingFinalize) {
-            return;
-        }
-        let content = json!({
-            "content": segment.buffer,
-            "status": "done",
-            "duration_ms": duration_ms,
-        })
-        .to_string();
-        let row = MessageRow {
-            id: segment.id.clone(),
-            conversation_id: self.conversation_id.clone(),
-            msg_id: Some(segment.id),
-            r#type: "thinking".into(),
-            content,
-            position: Some("left".into()),
-            status: Some("finish".into()),
-            hidden: false,
-            created_at: segment.started_at,
-        };
-        if let Err(e) = self.repo.insert_message(&row).await {
-            log_persist_error(&e, "Failed to persist thinking message");
-        }
+        self.adapter.persist_thinking_segment(segment, duration_ms as u64).await;
     }
 
     #[tracing::instrument(skip_all)]
@@ -758,234 +532,8 @@ impl StreamRelay {
         let Some(text_segment) = active_text.take() else {
             return;
         };
-        if let Some(segment) = self.finalize_text_segment(text_segment, status).await {
+        if let Some(segment) = self.adapter.finalize_text_segment(text_segment, status).await {
             text_segments.push(segment);
-        }
-    }
-
-    /// Persist a Gemini-style tool_call event.
-    #[tracing::instrument(skip_all)]
-    async fn persist_tool_call(&self, data: &aionui_ai_agent::protocol::events::tool_call::ToolCallEventData) {
-        if !self.allows_write(RuntimeWriteKind::ToolCallPersist) {
-            return;
-        }
-        if data.call_id.trim().is_empty() {
-            warn!(
-                tool = %data.name,
-                status = ?data.status,
-                "Skipping tool_call persistence because call_id is empty"
-            );
-            return;
-        }
-
-        let status = match data.status {
-            ToolCallStatus::Running => "work",
-            ToolCallStatus::Completed => "finish",
-            ToolCallStatus::Error => "error",
-        };
-        let content = serde_json::to_string(data).unwrap_or_default();
-
-        let existing = self
-            .repo
-            .get_message_by_msg_id(&self.conversation_id, &data.call_id, "tool_call")
-            .await
-            .unwrap_or(None);
-
-        if let Some(existing_row) = existing {
-            let merged_content = Self::merge_json_content(&existing_row.content, &content);
-            let update = aionui_db::MessageRowUpdate {
-                content: Some(merged_content),
-                status: Some(Some(status.to_owned())),
-                hidden: None,
-            };
-            if let Err(e) = self.repo.update_message(&data.call_id, &update).await {
-                error!(
-                    call_id = %data.call_id,
-                    tool = %data.name,
-                    status,
-                    error = %ErrorChain(&e),
-                    "Failed to update tool_call message"
-                );
-            } else {
-                debug!(
-                    call_id = %data.call_id,
-                    tool = %data.name,
-                    status,
-                    "Updated tool_call message"
-                );
-            }
-        } else {
-            let row = MessageRow {
-                id: data.call_id.clone(),
-                conversation_id: self.conversation_id.clone(),
-                msg_id: Some(data.call_id.clone()),
-                r#type: "tool_call".into(),
-                content,
-                position: Some("left".into()),
-                status: Some(status.to_owned()),
-                hidden: false,
-                created_at: now_ms(),
-            };
-            if let Err(e) = self.repo.insert_message(&row).await {
-                error!(
-                    call_id = %data.call_id,
-                    tool = %data.name,
-                    status,
-                    error = %ErrorChain(&e),
-                    "Failed to persist tool_call message"
-                );
-            } else {
-                debug!(
-                    call_id = %data.call_id,
-                    tool = %data.name,
-                    status,
-                    "Persisted tool_call message"
-                );
-            }
-        }
-    }
-
-    /// Persist an ACP (Claude CLI) tool call event.
-    /// First event (ToolCall) inserts; subsequent events (ToolCallUpdate) update.
-    #[tracing::instrument(skip_all)]
-    async fn persist_acp_tool_call(&self, data: &aionui_ai_agent::protocol::events::tool_call::AcpToolCallEventData) {
-        if !self.allows_write(RuntimeWriteKind::AcpToolCallPersist) {
-            return;
-        }
-        let tool_call_id = &data.update.tool_call_id;
-        let status = match data.update.status {
-            Some(AcpToolCallStatus::Pending) | None => "work",
-            Some(AcpToolCallStatus::InProgress) => "work",
-            Some(AcpToolCallStatus::Completed) => "finish",
-            Some(AcpToolCallStatus::Failed) => "error",
-        };
-
-        let mut value = serde_json::to_value(data).unwrap_or_default();
-        normalize_keys_to_snake_case(&mut value);
-        let content = value.to_string();
-
-        match data.update.session_update {
-            AcpToolCallSessionUpdateKind::ToolCall => {
-                let row = MessageRow {
-                    id: tool_call_id.clone(),
-                    conversation_id: self.conversation_id.clone(),
-                    msg_id: Some(tool_call_id.clone()),
-                    r#type: "acp_tool_call".into(),
-                    content,
-                    position: Some("left".into()),
-                    status: Some(status.to_owned()),
-                    hidden: false,
-                    created_at: now_ms(),
-                };
-                if let Err(e) = self.repo.insert_message(&row).await {
-                    error!(error = %ErrorChain(&e), "Failed to persist acp_tool_call message");
-                }
-            }
-            AcpToolCallSessionUpdateKind::ToolCallUpdate => {
-                let merged_content = self.merge_acp_tool_call_content(tool_call_id, &value).await;
-                let update = aionui_db::MessageRowUpdate {
-                    content: Some(merged_content),
-                    status: Some(Some(status.to_owned())),
-                    hidden: None,
-                };
-                if let Err(e) = self.repo.update_message(tool_call_id, &update).await {
-                    error!(error = %ErrorChain(&e), "Failed to update acp_tool_call message");
-                }
-            }
-        }
-    }
-
-    /// Merge two JSON content strings: overlays non-null fields from `new_json`
-    /// onto `existing_json`, preserving fields only present in the original.
-    fn merge_json_content(existing_json: &str, new_json: &str) -> String {
-        let mut base: serde_json::Value = serde_json::from_str(existing_json).unwrap_or_default();
-        let new_value: serde_json::Value = serde_json::from_str(new_json).unwrap_or_default();
-        if let (Some(base_obj), Some(new_obj)) = (base.as_object_mut(), new_value.as_object()) {
-            for (key, val) in new_obj {
-                if !val.is_null() {
-                    base_obj.insert(key.clone(), val.clone());
-                }
-            }
-        }
-        base.to_string()
-    }
-
-    /// Merge an AcpToolCall update into the existing DB record.
-    /// Reads the stored content, overlays non-null fields from the update,
-    /// preserving fields like `raw_input` that the update event omits.
-    async fn merge_acp_tool_call_content(&self, tool_call_id: &str, update_value: &serde_json::Value) -> String {
-        let existing = self
-            .repo
-            .get_message_by_msg_id(&self.conversation_id, tool_call_id, "acp_tool_call")
-            .await
-            .ok()
-            .flatten();
-
-        let Some(existing_row) = existing else {
-            return update_value.to_string();
-        };
-
-        let mut base: serde_json::Value = serde_json::from_str(&existing_row.content).unwrap_or_default();
-        if let (Some(base_update), Some(new_update)) = (
-            base.get_mut("update").and_then(|v| v.as_object_mut()),
-            update_value.get("update").and_then(|v| v.as_object()),
-        ) {
-            for (key, val) in new_update {
-                if !val.is_null() {
-                    base_update.insert(key.clone(), val.clone());
-                }
-            }
-        }
-        base.to_string()
-    }
-
-    /// Persist a tool_group event (array of tool summaries).
-    #[tracing::instrument(skip_all)]
-    async fn persist_tool_group(&self, entries: &[aionui_ai_agent::protocol::events::tool_call::ToolGroupEntry]) {
-        if !self.allows_write(RuntimeWriteKind::ToolGroupPersist) {
-            return;
-        }
-        let all_done = entries
-            .iter()
-            .all(|e| matches!(e.status, ToolCallStatus::Completed | ToolCallStatus::Error));
-        let status = if all_done { "finish" } else { "work" };
-        let content = serde_json::to_string(entries).unwrap_or_default();
-
-        let group_id = entries
-            .first()
-            .map(|e| e.call_id.clone())
-            .unwrap_or_else(ConversationService::mint_msg_id);
-
-        let existing = self
-            .repo
-            .get_message_by_msg_id(&self.conversation_id, &group_id, "tool_group")
-            .await
-            .unwrap_or(None);
-
-        if existing.is_some() {
-            let update = aionui_db::MessageRowUpdate {
-                content: Some(content),
-                status: Some(Some(status.to_owned())),
-                hidden: None,
-            };
-            if let Err(e) = self.repo.update_message(&group_id, &update).await {
-                error!(error = %ErrorChain(&e), "Failed to update tool_group message");
-            }
-        } else {
-            let row = MessageRow {
-                id: group_id.clone(),
-                conversation_id: self.conversation_id.clone(),
-                msg_id: Some(group_id),
-                r#type: "tool_group".into(),
-                content,
-                position: Some("left".into()),
-                status: Some(status.to_owned()),
-                hidden: false,
-                created_at: now_ms(),
-            };
-            if let Err(e) = self.repo.insert_message(&row).await {
-                error!(error = %ErrorChain(&e), "Failed to persist tool_group message");
-            }
         }
     }
 
@@ -1037,44 +585,6 @@ impl StreamRelay {
         let msg = WebSocketMessage::new("message.stream", payload);
         self.broadcaster.broadcast(msg);
     }
-
-    #[tracing::instrument(skip_all, fields(conversation_id = %conversation_id))]
-    pub async fn complete_conversation(
-        repo: &Arc<dyn IConversationRepository>,
-        broadcaster: &Arc<dyn EventBroadcaster>,
-        conversation_id: &str,
-    ) {
-        Self::complete_conversation_with_runtime(repo, broadcaster, conversation_id, None).await;
-    }
-
-    #[tracing::instrument(skip_all, fields(conversation_id = %conversation_id))]
-    pub async fn complete_conversation_with_runtime(
-        repo: &Arc<dyn IConversationRepository>,
-        broadcaster: &Arc<dyn EventBroadcaster>,
-        conversation_id: &str,
-        runtime: Option<ConversationRuntimeSummary>,
-    ) {
-        let update = aionui_db::ConversationRowUpdate {
-            status: Some("finished".to_owned()),
-            updated_at: Some(now_ms()),
-            ..Default::default()
-        };
-        if let Err(e) = repo.update(conversation_id, &update).await {
-            log_persist_error(&e, "Failed to update conversation status");
-        }
-
-        let payload = json!({
-            "conversation_id": conversation_id,
-            "session_id": conversation_id,
-            "status": "finished",
-            "canSendMessage": true,
-            "runtime": runtime,
-        });
-        let msg = WebSocketMessage::new("turn.completed", payload);
-        broadcaster.broadcast(msg);
-
-        debug!(conversation_id, status = "finished", "Turn completed");
-    }
 }
 
 struct SharedCronService(Arc<dyn ICronService>);
@@ -1111,9 +621,11 @@ impl ICronService for SharedCronService {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::stream_persistence::StreamPersistenceAdapter;
     use aionui_ai_agent::AgentError;
     use aionui_ai_agent::protocol::events::{ErrorEventData, FinishEventData, TextEventData, ThinkingEventData};
     use aionui_db::DbError;
+    use aionui_db::models::MessageRow;
     use std::sync::Mutex;
     use std::sync::atomic::{AtomicBool, Ordering};
 
@@ -1929,8 +1441,9 @@ mod tests {
         repo.set_not_found(true);
         let repo: Arc<dyn IConversationRepository> = repo;
         let bus: Arc<dyn EventBroadcaster> = Arc::new(aionui_realtime::BroadcastEventBus::new(64));
+        let adapter = StreamPersistenceAdapter::new("deleted-conv".into(), "msg-1".into(), repo, None);
 
-        StreamRelay::complete_conversation_with_runtime(&repo, &bus, "deleted-conv", None).await;
+        adapter.complete_conversation(&bus, None).await;
     }
 
     #[tokio::test]
